Fixes concerning dropdowns
