+++ src/shared/components/dropdown.js
@@ -1,76 +1,11 @@
 import DOMNavigator from "shared/dom-navigator.js";
-import { CustomElement } from './element.js';
 import { converse, api } from "@converse/headless/core";
 import { html } from 'lit';
 import { until } from 'lit/directives/until.js';
-
-const u = converse.env.utils;
-
-
-export class BaseDropdown extends CustomElement {
-
-    connectedCallback() {
-        super.connectedCallback();
-        this.registerEvents();
-    }
-
-    registerEvents() {
-        this.clickOutside = this._clickOutside.bind(this);
-        document.addEventListener('click', this.clickOutside);
-    }
-
-    firstUpdated () {
-        super.firstUpdated();
-        this.initArrowNavigation();
-        this.menu = this.querySelector('.dropdown-menu');
-        this.dropdown = this.firstElementChild;
-        this.button = this.dropdown.querySelector('button');
-        this.dropdown.addEventListener('click', ev => this.toggleMenu(ev));
-        this.dropdown.addEventListener('keyup', ev => this.handleKeyUp(ev));
-    }
-
-    _clickOutside(ev) {
-        if (!this.contains(ev.composedPath()[0])) {
-            this.hideMenu(ev);
-        }
-    }
-
-    hideMenu () {
-        u.removeClass('show', this.menu);
-        this.button?.setAttribute('aria-expanded', false);
-        this.button?.blur();
-    }
-
-    showMenu () {
-        u.addClass('show', this.menu);
-        this.button.setAttribute('aria-expanded', true);
-    }
-
-    toggleMenu (ev) {
-        ev.preventDefault();
-        if (u.hasClass('show', this.menu)) {
-            this.hideMenu();
-        } else {
-            this.showMenu();
-        }
-    }
-
-    handleKeyUp (ev) {
-        if (ev.keyCode === converse.keycodes.ESCAPE) {
-            this.hideMenu();
-        } else if (ev.keyCode === converse.keycodes.DOWN_ARROW && this.navigator && !this.navigator.enabled) {
-            this.enableArrowNavigation(ev);
-        }
-    }
-
-    disconnectedCallback () {
-        document.removeEventListener('click', this.clickOutside);
-        super.disconnectedCallback();
-    }
-}
+import DropdownBase from 'shared/components/dropdownbase.js';
 
 
-export default class DropdownList extends BaseDropdown {
+export default class Dropdown extends DropdownBase {
 
     static get properties () {
         return {
@@ -93,9 +28,14 @@ export default class DropdownList extends BaseDropdown {
         `;
     }
 
+    firstUpdated () {
+        super.firstUpdated();
+        this.initArrowNavigation();
+    }
+
     hideMenu () {
         super.hideMenu();
-        this.navigator.disable();
+        this.navigator?.disable();
     }
 
     initArrowNavigation () {
@@ -125,4 +65,4 @@ export default class DropdownList extends BaseDropdown {
     }
 }
 
-api.elements.define('converse-dropdown', DropdownList);
+api.elements.define('converse-dropdown', Dropdown);
